bookexchange/requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
Django==4.1.4
django-cors-headers==3.14.0
django-extra-fields==3.0.2
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.2.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.15
